travels/pyproject.toml
[tool.poetry]
name = "travels"
version = "0.1.0"
description = ""
authors = ["Pietro Dichiara <dichiarad@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.2"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
